sensor_id and type_id are now uint16_t

+++ Bouy-Firmware/include/sensor.h
@@ -12,7 +12,7 @@
  */
 class Sensor {
 public:
-  Sensor(uint32_t sensor_id, uint32_t type_id) : _sensor_id(sensor_id), _type_id(type_id) {}
+  Sensor(uint16_t sensor_id, uint16_t type_id) : _sensor_id(sensor_id), _type_id(type_id) {}
   virtual ~Sensor() = default;
 
   // METHODS
@@ -21,13 +21,13 @@ public:
   void disable();
 
   // GETTERS
-  uint32_t get_sensor_id();
-  uint32_t get_type_id();
+  uint16_t get_sensor_id();
+  uint16_t get_type_id();
   SensorStatus get_status();
 
 protected:
-  uint32_t _sensor_id;
-  uint32_t _type_id;
+  uint16_t _sensor_id;
+  uint16_t _type_id;
   SensorStatus _status;
 };
